organize donate files structure
