+++ src/components/views/donate/DonationCard.tsx
@@ -5,8 +5,8 @@ import { useIntl } from 'react-intl';
 import { useRouter } from 'next/router';
 import { isAddress } from 'viem';
 import { Shadow } from '@/components/styled-components/Shadow';
-import { RecurringDonationCard } from './RecurringDonationCard';
-import OneTimeDonationCard from './OneTimeDonationCard';
+import { RecurringDonationCard } from './Recurring/RecurringDonationCard';
+import OneTimeDonationCard from './OnTime/OneTimeDonationCard';
 import config from '@/configuration';
 import { useDonateData } from '@/context/donate.context';
 import { ChainType } from '@/types/config';
