+++ src/components/views/donate/SuccessView.tsx
@@ -29,7 +29,7 @@ import links from '@/lib/constants/links';
 import { EContentType } from '@/lib/constants/shareContent';
 import QFToast from './QFToast';
 import { DonationInfo } from './DonationInfo';
-import { ManageRecurringDonation } from './ManageRecurringDonation';
+import { ManageRecurringDonation } from './Recurring/ManageRecurringDonation';
 
 export const SuccessView: FC = () => {
 	const { formatMessage } = useIntl();
